refactor(detail): split RetrospectiveSection, ReviewSection, TimelineEvents into per-component files
Move shared types and helpers into lib/retrospective.ts and lib/review.ts.
Add TimelineContext to lib/timeline.ts. Break the three large section files
into focused per-component modules under retrospective/, review/, and
timeline/ subfolders. Top-level files drop from 2349 to 483 lines.
