+++ apps/web/src/components/detail/components/ReviewSection.tsx
@@ -1,105 +1,19 @@
 import { fetchEvents } from '@/lib/api';
 import type { AgentEventDto } from '@/lib/types';
 import { useQuery } from '@tanstack/react-query';
-import { AlertTriangle, Check, CheckCircle, Clock, GitPullRequest, XCircle } from 'lucide-react';
+import { Clock, GitPullRequest } from 'lucide-react';
 import { useIssueCostsBreakdown } from '../lib/costs';
+import { type ReviewPayload, VERDICT_LABEL } from '../lib/review';
 import { CostBadge } from './CostBadge';
 import { SectionEmptyState } from './SectionEmptyState';
+import { ChecklistRow } from './review/ChecklistRow';
+import { FindingsList } from './review/FindingsList';
+import { type PipelineStep, PreMergePipeline } from './review/PreMergePipeline';
+import { VerdictPill } from './review/VerdictPill';
 
 interface ReviewSectionProps {
   projectSlug: string;
   id: string;
-}
-
-interface CriterionCheck {
-  criterion: string;
-  status: 'met' | 'unmet' | 'unclear';
-  notes?: string;
-}
-
-type ReviewDisposition = 'fixed' | 'registered' | 'out-of-scope';
-
-interface ReviewFinding {
-  criterion?: string;
-  severity: 'blocker' | 'major' | 'minor';
-  description: string;
-  suggestion?: string;
-  file?: string;
-  line?: number;
-  disposition?: ReviewDisposition;
-  dispositionRef?: string;
-}
-
-const DISPOSITION_COLOR: Record<ReviewDisposition, string> = {
-  fixed: 'bg-green-500/15 text-green-400',
-  registered: 'bg-sky-500/15 text-sky-400',
-  'out-of-scope': 'bg-gray-500/15 text-gray-400',
-};
-
-function formatDisposition(d: ReviewDisposition, ref?: string): string {
-  if (d === 'fixed') return 'fixed';
-  if (d === 'out-of-scope') return 'out-of-scope';
-  return ref != null && ref.length > 0
-    ? `registered ${ref.startsWith('#') ? ref : `#${ref}`}`
-    : 'registered';
-}
-
-type ReviewVerdict = 'approved' | 'needs-fix' | 'needs-human';
-
-interface ReviewPayload {
-  verdict: ReviewVerdict;
-  confidence: number;
-  criteriaChecks: CriterionCheck[];
-  findings: ReviewFinding[];
-  escalationReason?: string;
-}
-
-const SEVERITY_COLOR: Record<string, string> = {
-  blocker: 'bg-red-500/15 text-red-400',
-  major: 'bg-orange-500/15 text-orange-400',
-  minor: 'bg-gray-500/15 text-gray-400',
-};
-
-const VERDICT_LABEL: Record<ReviewVerdict, string> = {
-  approved: 'Approved',
-  'needs-fix': 'Needs fix',
-  'needs-human': 'Needs human',
-};
-
-function ChecklistRow({ check, isFirst }: { check: CriterionCheck; isFirst: boolean }) {
-  const met = check.status === 'met';
-  const unclear = check.status === 'unclear';
-  const textColor = met ? 'text-fg' : unclear ? 'text-fg-2' : 'text-fg-3';
-
-  return (
-    <div
-      data-testid="review-checklist-row"
-      data-status={check.status}
-      className={`flex items-center gap-3 px-4 py-3 ${isFirst ? '' : 'border-t border-line'}`}
-    >
-      <span
-        aria-hidden
-        className={`grid place-items-center shrink-0 rounded-full ${
-          met
-            ? 'bg-[color:var(--success)]'
-            : unclear
-              ? 'border-[1.5px] border-dashed border-yellow-500/60'
-              : 'border-[1.5px] border-dashed border-line-2'
-        }`}
-        style={{ width: 22, height: 22 }}
-      >
-        {met && <Check size={12} strokeWidth={2.4} className="text-[color:var(--bg)]" />}
-        {check.status === 'unmet' && (
-          <XCircle size={12} className="text-red-400/70" strokeWidth={2} />
-        )}
-        {unclear && <AlertTriangle size={11} className="text-yellow-500/80" />}
-      </span>
-      <div className="flex-1 min-w-0">
-        <div className={`text-[13px] ${textColor}`}>{check.criterion}</div>
-        {check.notes && <div className="text-[11.5px] text-fg-3 mt-0.5">{check.notes}</div>}
-      </div>
-    </div>
-  );
 }
 
 export function ReviewSection({ projectSlug, id }: ReviewSectionProps) {
@@ -148,7 +62,7 @@ export function ReviewSection({ projectSlug, id }: ReviewSectionProps) {
       }
     | undefined;
 
-  const pipeline: { label: string; passed: boolean | undefined }[] = [
+  const pipeline: PipelineStep[] = [
     { label: 'Lint clean', passed: qaPayload?.tierResults?.structural?.passed },
     {
       label: 'Tests passing',
@@ -171,19 +85,6 @@ export function ReviewSection({ projectSlug, id }: ReviewSectionProps) {
   const checks = review.criteriaChecks ?? [];
   const metCount = checks.filter((c) => c.status === 'met').length;
   const total = checks.length;
-
-  const VerdictIcon =
-    review.verdict === 'approved'
-      ? CheckCircle
-      : review.verdict === 'needs-fix'
-        ? XCircle
-        : AlertTriangle;
-  const verdictColor =
-    review.verdict === 'approved'
-      ? 'text-[color:var(--success)]'
-      : review.verdict === 'needs-fix'
-        ? 'text-orange-500'
-        : 'text-red-500';
 
   const confidencePct = Math.round(review.confidence * 100);
   const verdictLabel = VERDICT_LABEL[review.verdict] ?? String(review.verdict ?? 'unknown');
@@ -230,25 +131,8 @@ export function ReviewSection({ projectSlug, id }: ReviewSectionProps) {
         )}
       </div>
 
-      {/* Verdict pill */}
-      <div className="flex items-center gap-2.5">
-        <span
-          data-testid="review-verdict-pill"
-          className={`inline-flex items-center gap-1.5 h-6 px-2.5 rounded-full border text-[11.5px] font-medium uppercase tracking-wide ${
-            review.verdict === 'approved'
-              ? 'border-[color:var(--success)]/40 bg-[color:var(--success)]/10 text-[color:var(--success)]'
-              : review.verdict === 'needs-fix'
-                ? 'border-orange-500/40 bg-orange-500/10 text-orange-400'
-                : 'border-red-500/40 bg-red-500/10 text-red-400'
-          }`}
-        >
-          <VerdictIcon size={11} className={verdictColor} />
-          {verdictLabel}
-        </span>
-        <span className="text-[11.5px] text-fg-3">Confidence {confidencePct}%</span>
-      </div>
+      <VerdictPill verdict={review.verdict} confidence={review.confidence} />
 
-      {/* Escalation reason */}
       {review.escalationReason && (
         <div className="p-3 bg-red-500/10 border border-red-500/20 rounded-lg text-[12.5px] text-fg-2">
           <span className="font-medium text-red-400">Escalation: </span>
@@ -256,50 +140,8 @@ export function ReviewSection({ projectSlug, id }: ReviewSectionProps) {
         </div>
       )}
 
-      {/* Pre-merge pipeline */}
-      <div>
-        <h3 className="text-[11px] font-medium text-fg-3 uppercase tracking-[0.14em] mb-2">
-          Pre-merge pipeline
-        </h3>
-        <div className="rounded-lg border border-line bg-bg-elev overflow-hidden">
-          {pipeline.map((p, i) => {
-            const passed = p.passed;
-            const unknown = passed === undefined;
-            return (
-              <div
-                key={p.label}
-                className={`flex items-center gap-3 px-4 py-3 ${i === 0 ? '' : 'border-t border-line'}`}
-              >
-                <span
-                  aria-hidden
-                  className={`grid place-items-center shrink-0 rounded-full ${
-                    passed === true
-                      ? 'bg-[color:var(--success)]'
-                      : unknown
-                        ? 'border-[1.5px] border-dashed border-line-2'
-                        : 'border-[1.5px] border-dashed border-yellow-500/60'
-                  }`}
-                  style={{ width: 22, height: 22 }}
-                >
-                  {passed === true && (
-                    <Check size={12} strokeWidth={2.4} className="text-[color:var(--bg)]" />
-                  )}
-                  {passed === false && (
-                    <XCircle size={12} className="text-red-400/70" strokeWidth={2} />
-                  )}
-                </span>
-                <span
-                  className={`text-[13px] ${passed === true ? 'text-fg' : unknown ? 'text-fg-3' : 'text-fg-2'}`}
-                >
-                  {p.label}
-                </span>
-              </div>
-            );
-          })}
-        </div>
-      </div>
+      <PreMergePipeline pipeline={pipeline} />
 
-      {/* Criteria checklist */}
       {total > 0 ? (
         <div className="rounded-lg border border-line bg-bg-elev overflow-hidden">
           {checks.map((c, i) => (
@@ -312,43 +154,7 @@ export function ReviewSection({ projectSlug, id }: ReviewSectionProps) {
         </div>
       )}
 
-      {/* Findings */}
-      {(review.findings?.length ?? 0) > 0 && (
-        <div>
-          <h3 className="text-[11px] font-medium text-fg-3 uppercase tracking-[0.14em] mb-3">
-            Findings
-          </h3>
-          <div className="space-y-2">
-            {review.findings.map((f) => (
-              <div key={f.description} className="border border-line rounded-lg p-3 text-[12.5px]">
-                <div className="flex items-center gap-2 mb-1 flex-wrap">
-                  <span
-                    className={`text-[10px] font-medium uppercase px-1.5 py-0.5 rounded ${SEVERITY_COLOR[f.severity] ?? 'bg-gray-500/15 text-gray-400'}`}
-                  >
-                    {f.severity}
-                  </span>
-                  {f.disposition != null && (
-                    <span
-                      className={`text-[10px] font-medium px-1.5 py-0.5 rounded ${DISPOSITION_COLOR[f.disposition]}`}
-                      data-testid="review-finding-disposition"
-                    >
-                      {formatDisposition(f.disposition, f.dispositionRef)}
-                    </span>
-                  )}
-                  {f.file && (
-                    <span className="font-mono text-[10px] text-fg-3 bg-bg-elev-2 px-1 py-0.5 rounded">
-                      {f.file}
-                      {f.line != null ? `:${f.line}` : ''}
-                    </span>
-                  )}
-                </div>
-                <div className="text-fg-2">{f.description}</div>
-                {f.suggestion && <div className="text-[11px] text-fg-3 mt-1">→ {f.suggestion}</div>}
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
+      <FindingsList findings={review.findings ?? []} />
     </div>
   );
 }
